+++ holidays_view.php
@@ -1,5 +1,10 @@
 <?php 	
 include('header.php'); 
+$user = getPermissions('holidays');
+if($user['view'] != '1')
+{
+	header("location:page_not_found.php");
+}
 $obj = new Holidays();
 $result = $obj->viewHoliday();
 ?>
@@ -23,7 +28,11 @@ $(document).ready(function(){
 		<?php if(isset($_SESSION['succ_msg'])){ echo $_SESSION['succ_msg']; unset($_SESSION['succ_msg']);} ?>
 		</div>
         <div class="full_w">
-            <div class="h_title">Holidays View<a href="holidays.php" class="gird-addnew" title="Add New Holiday">Add New</a></div>
+            <div class="h_title">Holidays View
+			<?php if($user['add_role'] != '0'){?>
+			<a href="holidays.php" class="gird-addnew" title="Add New Holiday">Add New</a>
+			<?php } ?>
+			</div>
             <table id="datatables" class="display">
                 <thead>
                     <tr>
@@ -32,7 +41,9 @@ $(document).ready(function(){
                         <th >Holiday Reason</th>
 						<th >Add Date</th>
 						<th >Update Date</th>
-                        <th >Action</th>
+                        <?php if($user['edit'] != '0' || $user['delete_role'] != '0'){?>
+								<th>Action</th>
+						<?php } ?>
                     </tr>
                 </thead>
                 <tbody>
@@ -44,10 +55,16 @@ $(document).ready(function(){
 						<td><?php echo $data['holiday_reason'] ?></td>
                         <td><?php echo $data['date_add'] ?></td>
 						<td><?php echo $data['date_update'] ?></td>
-                        <td class="align-center" id="<?php echo $data['id'] ?>">
-                            <a href="holidays.php?edit=<?php echo base64_encode($data['id']) ?>" class="table-icon edit" title="Edit"></a>
-							<a href="#" class="table-icon delete" onClick="deleteHoliday(<?php echo $data['id'] ?>)"></a>
-                        </td>
+						<?php if($user['edit'] != '0' || $user['delete_role'] != '0'){?>
+							<td class="align-center" id="<?php echo $data['id'] ?>">
+								<?php if($user['edit'] != '0'){?>
+									<a href="holidays.php?edit=<?php echo base64_encode($data['id']) ?>" class="table-icon edit" title="Edit"></a>
+								<?php } ?>
+								<?php if($user['delete_role'] != '0'){?>
+									<a href="#" class="table-icon delete" onClick="deleteHoliday(<?php echo $data['id'] ?>)"></a>
+								<?php } ?>
+							</td>
+						<?php } ?>
                     </tr>
 					<?php }?>
                 </tbody>
