+++ rooms_view.php
@@ -1,5 +1,10 @@
 <?php
 include('header.php');
+$user = getPermissions('rooms');
+if($user['view'] != '1')
+{
+	header("location:page_not_found.php");
+}
 $obj = new Classroom();
 $result = $obj->viewRoom();
 ?>
@@ -23,7 +28,11 @@ $(document).ready(function(){
 				<?php if(isset($_SESSION['succ_msg'])){ echo $_SESSION['succ_msg']; unset($_SESSION['succ_msg']);} ?>
 			</div>
 			<div class="full_w">
-				<div class="h_title">Classrooms View<a href="rooms.php" class="gird-addnew" title="Add New Classroom">Add new</a></div>
+				<div class="h_title">Classrooms View
+				<?php if($user['add_role'] != '0'){?>
+				<a href="rooms.php" class="gird-addnew" title="Add New Classroom">Add new</a>
+				<?php } ?>
+				</div>
 				<table id="datatables" class="display">
 					<thead>
 						<tr>
@@ -31,7 +40,9 @@ $(document).ready(function(){
 							<th>Name</th>
 							<th>Room Type</th>
 							<th>Building</th>
-							<th>Action</th>
+							<?php if($user['edit'] != '0' || $user['delete_role'] != '0'){?>
+								<th>Action</th>
+							<?php } ?>
 						</tr>
 					</thead>
 					<tbody>
@@ -41,10 +52,16 @@ $(document).ready(function(){
                         <td><?php echo $data['room_name'] ?></td>
                         <td><?php echo $data['room_type'] ?></td>
 						<td><?php echo $data['building_name'] ?></td>
-                        <td class="align-center" id="<?php echo $data['listId'] ?>">
-                            <a href="rooms.php?edit=<?php echo base64_encode($data['listId']) ?>" class="table-icon edit" title="Edit"></a>
-							<a href="#" class="table-icon delete" onClick="deleteRoom(<?php echo $data['listId'] ?>)"></a>
-                        </td>
+						<?php if($user['edit'] != '0' || $user['delete_role'] != '0'){?>
+							<td class="align-center" id="<?php echo $data['listId'] ?>">
+							<?php if($user['edit'] != '0'){?>
+								<a href="rooms.php?edit=<?php echo base64_encode($data['listId']) ?>" class="table-icon edit" title="Edit"></a>
+							<?php } ?>
+							<?php if($user['delete_role'] != '0'){?>
+								<a href="#" class="table-icon delete" onClick="deleteRoom(<?php echo $data['listId'] ?>)"></a>
+							<?php } ?>
+							</td>
+						<?php } ?>
                     </tr>
 				<?php }?>
 					</tbody>
