+++ special_activity_view.php
@@ -1,5 +1,10 @@
 <?php
 include('header.php');
+$user = getPermissions('special_activity');
+if($user['view'] != '1')
+{
+	header("location:page_not_found.php");
+}
 $obj=new SpecialActivity();
 $objT = new Teacher();
 $result=$obj->getSpecialActivityDetailView();
@@ -22,7 +27,11 @@ $(document).ready(function(){
     <div id="main">
 	<?php if(isset($_SESSION['succ_msg'])){ echo '<div class="full_w green center">'.$_SESSION['succ_msg'].'</div>'; unset($_SESSION['succ_msg']);} ?>
         <div class="full_w">
-            <div class="h_title">Special Activity View<a href="special_activity.php" class="gird-addnew" title="Add New Special Activity"> Add New</a></div>
+            <div class="h_title">Special Activity View
+			<?php if($user['add_role'] != '0'){?>
+			<a href="special_activity.php" class="gird-addnew" title="Add New Special Activity"> Add New</a>
+			<?php } ?>
+			</div>
             <table id="datatables" class="display">
                 <thead>
                     <tr>
@@ -35,7 +44,9 @@ $(document).ready(function(){
 						<th >Classroom</th>
                     	<th >Subject</th>
                         <th >Teacher</th>
-                        <th width="75" >Action</th>
+						<?php if($user['edit'] != '0' || $user['delete_role'] != '0'){?>
+								<th width="75" >Action</th>
+						<?php } ?>                        
                     </tr>
                 </thead>
                 <tbody>
@@ -105,10 +116,16 @@ $(document).ready(function(){
 						<td><?php if($data['room_id']==0){echo 'N/A';}else{echo $data['room_name'];}?></td>
 						<td><?php if($data['subject_id']==0){echo 'N/A';}else{echo $data['subject_name'];}?></td>
 						<td><?php if($data['teacher_id']==0){echo 'N/A';}else{echo $data['teacher_name'];}?></td>
-						<td id="<?php echo trim($data['special_activity_name']) ?>">
-							<a href="special_activity.php?gp_Edit=<?php echo base64_encode($data['special_activity_name'])?>" class="table-icon edit" title="Edit"></a>
-							<a class="table-icon delete" onClick="deleteSpecialActivity('<?php echo trim($data['special_activity_name']) ?>')"></a>
-						</td>
+						<?php if($user['edit'] != '0' || $user['delete_role'] != '0'){?>
+							<td id="<?php echo trim($data['special_activity_name']) ?>">
+								<?php if($user['edit'] != '0'){?>
+									<a href="special_activity.php?gp_Edit=<?php echo base64_encode($data['special_activity_name'])?>" class="table-icon edit" title="Edit"></a>
+								<?php } ?>
+								<?php if($user['delete_role'] != '0'){?>
+									<a class="table-icon delete" onClick="deleteSpecialActivity('<?php echo trim($data['special_activity_name']) ?>')"></a>
+								<?php } ?>
+							</td>
+						<?php } ?>
                     </tr>
 					<?php $i++;}?>
                 </tbody>
